+++ src/Reader/AddChildReader.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace KunicMarko\SonataAnnotationBundle\Reader;
 
-use KunicMarko\SonataAnnotationBundle\Annotation\ChildAdmin;
+use KunicMarko\SonataAnnotationBundle\Annotation\AddChild;
 
 /**
  * @author Marko Kunic <kunicmarko20@gmail.com>
  */
-final class ChildAdminReader
+final class AddChildReader
 {
     use AnnotationReaderTrait;
 
@@ -18,7 +18,7 @@ final class ChildAdminReader
         $children = [];
 
         foreach ($this->getClassAnnotations($class) as $annotation) {
-            if (!$annotation instanceof ChildAdmin) {
+            if (!$annotation instanceof AddChild) {
                 continue;
             }
