+++ src/Resources/config/reader.xml
@@ -13,6 +13,6 @@
         <service id="sonata.annotation.reader.dashboard_action" class="KunicMarko\SonataAnnotationBundle\Reader\DashboardActionReader" />
         <service id="sonata.annotation.reader.export" class="KunicMarko\SonataAnnotationBundle\Reader\ExportReader" />
         <service id="sonata.annotation.reader.datagrid_values" class="KunicMarko\SonataAnnotationBundle\Reader\DatagridValuesReader" />
-        <service id="sonata.annotation.reader.child_admin" class="KunicMarko\SonataAnnotationBundle\Reader\ChildAdminReader" />
+        <service id="sonata.annotation.reader.add_child" class="KunicMarko\SonataAnnotationBundle\Reader\AddChildReader"/>
     </services>
 </container>
