+++ tests/SonataAnnotationBundleTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace KunicMarko\SonataAnnotationBundle\Tests;
 
 use KunicMarko\SonataAnnotationBundle\DependencyInjection\Compiler\AccessCompilerPass;
-use KunicMarko\SonataAnnotationBundle\DependencyInjection\Compiler\AddChildrenCompilerPass;
+use KunicMarko\SonataAnnotationBundle\DependencyInjection\Compiler\AddChildCompilerPass;
 use KunicMarko\SonataAnnotationBundle\DependencyInjection\Compiler\AutoRegisterCompilerPass;
 use KunicMarko\SonataAnnotationBundle\SonataAnnotationBundle;
 use PHPUnit\Framework\TestCase;
@@ -48,7 +48,7 @@ final class SonataAnnotationBundleTest extends TestCase
         )->shouldBeCalledTimes(1);
 
         $containerBuilder->addCompilerPass(
-            Argument::type(AddChildrenCompilerPass::class),
+            Argument::type(AddChildCompilerPass::class),
             Argument::cetera()
         )->shouldBeCalledTimes(1);
